requirements.txt
torch>=2.0
fair-esm
scikit-learn
matplotlib
seaborn
numpy
pandas
